refactor: misc renames
