+++ psycop/common/feature_generation/sequences/event_loader.py
@@ -1,29 +1,20 @@
-import datetime
 from abc import ABC, abstractmethod
-from collections.abc import Sequence
-from typing import Union
+from typing import Protocol
 
 import polars as pl
 
-from psycop.common.data_structures.patient import Patient
-from psycop.common.feature_generation.loaders.raw.load_demographic import birthdays
-from psycop.common.feature_generation.loaders.raw.load_ids import SplitName, load_ids
-from psycop.common.feature_generation.loaders.raw.sql_load import sql_load
-from psycop.common.feature_generation.sequences.event_dataframes_to_patient import (
-    EventDataFramesToPatientSlices,
-)
-
 from ...sequence_models.registry import Registry
+from ..loaders.raw.sql_load import sql_load
 
 
-class EventDfLoader(ABC):
+class EventLoader(Protocol):
     @abstractmethod
     def load_events(self) -> pl.LazyFrame:
         ...
 
 
 @Registry.event_loaders.register("diagnoses")
-class DiagnosisLoader(EventDfLoader):
+class DiagnosisLoader(EventLoader):
     """Load all diagnoses for all patients."""
 
     def __init__(self) -> None:
@@ -75,61 +66,3 @@ class DiagnosisLoader(EventDfLoader):
         ).unique(maintain_order=True)
 
         return df.drop_nulls()  # Drop all rows with no diagnosis
-
-
-def keep_if_min_n_visits(
-    df: pl.LazyFrame,
-    n_visits: Union[int, None],
-) -> pl.LazyFrame:
-    df = df.filter(
-        pl.col("timestamp").unique().count().over("dw_ek_borger") >= n_visits,
-    )
-
-    return df
-
-
-class PatientLoader:
-    @staticmethod
-    def load_date_of_birth_df() -> pl.DataFrame:
-        df = pl.from_pandas(birthdays()).rename({"date_of_birth": "timestamp"})
-
-        return df
-
-    @staticmethod
-    def get_split(
-        event_loaders: Sequence[EventDfLoader],
-        split: SplitName,
-        min_n_events: int | None = None,
-        fraction: float = 1.0,
-    ) -> Sequence[Patient]:
-        event_data = pl.concat([loader.load_events() for loader in event_loaders])
-        split_ids = (
-            pl.from_pandas(load_ids(split=split)).sample(fraction=fraction).lazy()
-        )
-
-        events_from_train = split_ids.join(event_data, on="dw_ek_borger", how="left")
-
-        if min_n_events:
-            events_from_train = keep_if_min_n_visits(
-                events_from_train,
-                n_visits=min_n_events,
-            )
-
-        events_after_2013 = events_from_train.filter(
-            pl.col("timestamp") > datetime.datetime(2013, 1, 1),
-        )
-
-        unpacked_patients = EventDataFramesToPatientSlices().unpack(
-            source_event_dataframes=[events_after_2013.collect()],
-            date_of_birth_df=PatientLoader.load_date_of_birth_df(),
-        )
-
-        return unpacked_patients
-
-
-if __name__ == "__main__":
-    patients = PatientLoader.get_split(
-        event_loaders=[DiagnosisLoader()],
-        split=SplitName.TRAIN,
-        min_n_events=5,
-    )
