+++ psycop/common/feature_generation/sequences/test_patient_slice_from_event_df.py
@@ -5,12 +5,12 @@ import pytest
 from psycop.common.data_structures.static_feature import StaticFeature
 from psycop.common.data_structures.temporal_event import TemporalEvent
 from psycop.common.data_structures.test_patient import get_test_patient
-from psycop.common.feature_generation.sequences.event_dataframes_to_patient import (
-    EventDataFramesToPatientSlices,
-    PatientSliceColumnNames,
-)
-from psycop.common.feature_generation.sequences.patient_loaders import (
+from psycop.common.feature_generation.sequences.event_loader import (
     DiagnosisLoader,
+)
+from psycop.common.feature_generation.sequences.patient_slice_from_events import (
+    PatientSliceColumnNames,
+    PatientSliceFromEvents,
 )
 from psycop.common.feature_generation.sequences.utils_for_testing import (
     get_test_date_of_birth_df,
@@ -62,7 +62,7 @@ def test_temporal_events():
     )
     expected_patients = [patient_1, patient_2]
 
-    unpacked = EventDataFramesToPatientSlices(
+    unpacked = PatientSliceFromEvents(
         column_names=PatientSliceColumnNames(source_subtype_col_name=None),
     ).unpack(
         source_event_dataframes=[test_data],
@@ -84,7 +84,7 @@ def test_static_features():
         [StaticFeature(source_type="test", value=0)],
     )
 
-    unpacked = EventDataFramesToPatientSlices().unpack(
+    unpacked = PatientSliceFromEvents().unpack(
         source_event_dataframes=[test_data],
         date_of_birth_df=get_test_date_of_birth_df(patient_ids=[1]),
     )
@@ -117,7 +117,7 @@ def test_multiple_event_sources():
         [expected_static_event, expected_temporal_event],
     )
 
-    unpacked = EventDataFramesToPatientSlices(
+    unpacked = PatientSliceFromEvents(
         column_names=PatientSliceColumnNames(source_subtype_col_name=None),
     ).unpack(
         source_event_dataframes=[test_data, test_data2],
@@ -135,7 +135,7 @@ def test_patient_without_date_of_birth_raises_error():
     )
 
     with pytest.raises(KeyError):
-        EventDataFramesToPatientSlices().unpack(
+        PatientSliceFromEvents().unpack(
             source_event_dataframes=[test_data],
             date_of_birth_df=get_test_date_of_birth_df(patient_ids=[2]),
         )
@@ -156,14 +156,14 @@ def test_passing_patient_colnames():
         DiagnosisLoader().preprocess_diagnosis_columns(df=df.lazy()).collect()
     )
 
-    unpacked_with_source_subtype_column = EventDataFramesToPatientSlices(
+    unpacked_with_source_subtype_column = PatientSliceFromEvents(
         column_names=PatientSliceColumnNames(source_subtype_col_name="type"),
     ).unpack(
         source_event_dataframes=[formatted_df],
         date_of_birth_df=get_test_date_of_birth_df(patient_ids=[1, 2]),
     )
 
-    unpacked_without_source_subtype_column = EventDataFramesToPatientSlices(
+    unpacked_without_source_subtype_column = PatientSliceFromEvents(
         column_names=PatientSliceColumnNames(source_subtype_col_name=None),
     ).unpack(
         source_event_dataframes=[formatted_df],
